#130 first stab at app restructure with app-starter dir
