+++ app-starter/.gitignore
@@ -1,4 +1,5 @@
 # Ignore everything in this directory
-*
+#*
 # Except this file
-!.gitignore
+#!.gitignore
+
