+++ app-starter/components/AppFooter.vue
