+++ app-starter/static/app-conf-minimal.json
