+++ app-starter/static/app-conf-projected.json
