+++ app-starter/static/app-conf.json
