+++ app-starter/static/css/app.css
@@ -1,0 +1,1 @@
+/* Created by init-app.js at Thu Aug 08 2019 15:59:47 GMT+0200 (GMT+02:00) */
