+++ app-starter/static/data/shops-dannstadt.geojson
