+++ app/README.md
@@ -1,0 +1,8 @@
+# Wegue App Dir
+This directory will contain your Wegue Application.
+By using the command `npm run init:app` this dir is populated with
+the contents from the `starter-app` dir.
+
+You can then place custom components under `components` here.
+Static content like data, icons and css can be placed under static.
+The default config file is `app-conf.json`.
