+++ package.json
@@ -25,8 +25,8 @@
     "vuetify": "1.5.23"
   },
   "devDependencies": {
-    "@vue/test-utils": "^1.0.0-beta.29",
-    "autoprefixer": "^9.6.0",
+    "@vue/test-utils": "^1.0.3",
+    "autoprefixer": "^9.8.0",
     "babel-core": "^6.22.1",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^7.1.1",
@@ -38,17 +38,18 @@
     "chai": "^4.2.0",
     "chalk": "^2.4.2",
     "connect-history-api-fallback": "^1.6.0",
+    "copy-dir": "^1.2.0",
     "copy-webpack-plugin": "^4.3.1",
-    "cross-env": "^5.1.3",
+    "cross-env": "^5.2.1",
     "css-loader": "^1.0.1",
     "cssnano": "^3.10.0",
     "eslint": "^5.16.0",
     "eslint-config-standard": "^6.2.1",
     "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^2.1.2",
+    "eslint-loader": "^2.2.1",
     "eslint-plugin-html": "^4.0.6",
-    "eslint-plugin-promise": "^4.1.1",
-    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.17.1",
     "extract-text-webpack-plugin": "^2.0.0",
@@ -56,9 +57,9 @@
     "friendly-errors-webpack-plugin": "^1.1.3",
     "gh-pages": "^1.1.0",
     "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.19.1",
+    "http-proxy-middleware": "^0.19.2",
     "inject-loader": "^4.0.1",
-    "karma": "^4.1.0",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-coverage": "^1.1.1",
     "karma-mocha": "^1.3.0",
@@ -68,25 +69,25 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "^0.0.32",
     "karma-webpack": "^3.0.5",
-    "mocha": "^6.1.0",
+    "mocha": "^6.2.3",
     "opn": "^6.0.0",
     "optimize-css-assets-webpack-plugin": "^2.0.0",
     "ora": "^3.4.0",
     "phantomjs-prebuilt": "^2.1.16",
-    "rimraf": "^2.6.3",
-    "semver": "^6.1.1",
-    "shelljs": "^0.8.3",
-    "sinon": "^7.3.0",
-    "sinon-chai": "^3.3.0",
+    "rimraf": "^2.7.1",
+    "semver": "^6.3.0",
+    "shelljs": "^0.8.4",
+    "sinon": "^7.5.0",
+    "sinon-chai": "^3.5.0",
     "url-loader": "^1.1.2",
     "vue-loader": "^14.2.4",
     "vue-style-loader": "^4.1.2",
     "vue-template-compiler": "2.6.11",
     "webpack": "^2.6.1",
-    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-bundle-analyzer": "^3.8.0",
     "webpack-dev-middleware": "^1.12.2",
     "webpack-hot-middleware": "^2.25.0",
-    "webpack-merge": "^4.2.1"
+    "webpack-merge": "^4.2.2"
   },
   "engines": {
     "node": ">= 4.0.0",
